+++ sdk/sdk.go
@@ -22,6 +22,7 @@ import (
 	"github.com/opentdf/platform/protocol/go/policy/resourcemapping"
 	"github.com/opentdf/platform/protocol/go/policy/subjectmapping"
 	"github.com/opentdf/platform/protocol/go/wellknownconfiguration"
+	"github.com/opentdf/platform/sdk/audit"
 	"github.com/opentdf/platform/sdk/auth"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
@@ -41,9 +42,9 @@ func (c Error) Error() string {
 }
 
 type SDK struct {
+	config
 	conn                    *grpc.ClientConn
 	dialOptions             []grpc.DialOption
-	kasSessionKey           ocrypto.RsaKeyPair
 	tokenSource             auth.AccessTokenSource
 	Namespaces              namespaces.NamespaceServiceClient
 	Attributes              attributes.AttributesServiceClient
@@ -52,7 +53,6 @@ type SDK struct {
 	KeyAccessServerRegistry kasregistry.KeyAccessServerRegistryServiceClient
 	Authorization           authorization.AuthorizationServiceClient
 	EntityResoution         entityresolution.EntityResolutionServiceClient
-	platformConfiguration   PlatformConfiguration
 	wellknownConfiguration  wellknownconfiguration.WellKnownServiceClient
 }
 
@@ -120,14 +120,21 @@ func New(platformEndpoint string, opts ...Option) (*SDK, error) {
 		}
 	}
 
+	var uci []grpc.UnaryClientInterceptor
+
 	accessTokenSource, err := buildIDPTokenSource(cfg)
 	if err != nil {
 		return nil, err
 	}
 	if accessTokenSource != nil {
 		interceptor := auth.NewTokenAddingInterceptor(accessTokenSource, cfg.tlsConfig)
-		dialOptions = append(dialOptions, grpc.WithUnaryInterceptor(interceptor.AddCredentials))
+		uci = append(uci, interceptor.AddCredentials)
 	}
+
+	// Add request ID interceptor
+	uci = append(uci, audit.MetadataAddingClientInterceptor)
+
+	dialOptions = append(dialOptions, grpc.WithChainUnaryInterceptor(uci...))
 
 	if platformEndpoint != "" {
 		var err error
@@ -138,11 +145,10 @@ func New(platformEndpoint string, opts ...Option) (*SDK, error) {
 	}
 
 	return &SDK{
+		config:                  *cfg,
 		conn:                    defaultConn,
 		dialOptions:             dialOptions,
 		tokenSource:             accessTokenSource,
-		kasSessionKey:           *cfg.kasSessionKey,
-		platformConfiguration:   cfg.platformConfiguration,
 		Attributes:              attributes.NewAttributesServiceClient(selectConn(cfg.policyConn, defaultConn)),
 		Namespaces:              namespaces.NewNamespaceServiceClient(selectConn(cfg.policyConn, defaultConn)),
 		ResourceMapping:         resourcemapping.NewResourceMappingServiceClient(selectConn(cfg.policyConn, defaultConn)),