+++ service/internal/auth/authn.go
@@ -5,20 +5,23 @@ import (
 	"crypto"
 	"crypto/sha256"
 	"encoding/base64"
+	"encoding/json"
 	"fmt"
 	"log/slog"
 	"net/http"
 	"net/url"
-	"path/filepath"
 	"slices"
 	"strings"
 	"time"
 
+	"github.com/bmatcuk/doublestar"
 	"github.com/lestrrat-go/jwx/v2/jwa"
 	"github.com/lestrrat-go/jwx/v2/jwk"
 	"github.com/lestrrat-go/jwx/v2/jws"
 	"github.com/lestrrat-go/jwx/v2/jwt"
-	"github.com/opentdf/platform/service/internal/logger/audit"
+
+	sdkAudit "github.com/opentdf/platform/sdk/audit"
+	"github.com/opentdf/platform/service/internal/logger"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
@@ -71,29 +74,31 @@ const refreshInterval = 15 * time.Minute
 // Authentication holds a jwks cache and information about the openid configuration
 type Authentication struct {
 	enforceDPoP bool
-	// cache holds the jwks cache
-	cache *jwk.Cache
-	// openidConfigurations holds the openid configuration for each issuer
-	oidcConfigurations map[string]AuthNConfig
+	// keySet holds a cached key set
+	cachedKeySet jwk.Set
+	// openidConfigurations holds the openid configuration for the issuer
+	oidcConfiguration AuthNConfig
 	// Casbin enforcer
 	enforcer *Enforcer
 	// Public Routes HTTP & gRPC
 	publicRoutes []string
+	// Custom Logger
+	logger *logger.Logger
 }
 
 // Creates new authN which is used to verify tokens for a set of given issuers
-func NewAuthenticator(ctx context.Context, cfg Config) (*Authentication, error) {
+func NewAuthenticator(ctx context.Context, cfg Config, logr *logger.Logger, wellknownRegistration func(namespace string, config any) error) (*Authentication, error) {
 	a := &Authentication{
 		enforceDPoP: cfg.EnforceDPoP,
+		logger:      logr,
 	}
-	a.oidcConfigurations = make(map[string]AuthNConfig)
 
 	// validate the configuration
 	if err := cfg.validateAuthNConfig(); err != nil {
 		return nil, err
 	}
 
-	a.cache = jwk.NewCache(ctx)
+	cache := jwk.NewCache(ctx)
 
 	// Build new cache
 	// Discover OIDC Configuration
@@ -102,38 +107,67 @@ func NewAuthenticator(ctx context.Context, cfg Config) (*Authentication, error) 
 		return nil, err
 	}
 
-	cfg.OIDCConfiguration = *oidcConfig
+	// If the issuer is different from the one in the configuration, update the configuration
+	// This could happen if we are hitting an internal endpoint. Example we might point to https://keycloak.opentdf.svc/realms/opentdf
+	// but the external facing issuer is https://keycloak.opentdf.local/realms/opentdf
+	if oidcConfig.Issuer != cfg.Issuer {
+		cfg.Issuer = oidcConfig.Issuer
+	}
 
 	cacheInterval, err := time.ParseDuration(cfg.CacheRefresh)
 	if err != nil {
-		slog.ErrorContext(ctx, fmt.Sprintf("Invalid cache_refresh_interval [%s]", cfg.CacheRefresh), "err", err)
+		logr.ErrorContext(ctx, fmt.Sprintf("Invalid cache_refresh_interval [%s]", cfg.CacheRefresh), "err", err)
 		cacheInterval = refreshInterval
 	}
 
 	// Register the jwks_uri with the cache
-	if err := a.cache.Register(cfg.JwksURI, jwk.WithMinRefreshInterval(cacheInterval)); err != nil {
+	if err := cache.Register(oidcConfig.JwksURI, jwk.WithMinRefreshInterval(cacheInterval)); err != nil {
 		return nil, err
 	}
 
 	casbinConfig := CasbinConfig{
 		PolicyConfig: cfg.Policy,
 	}
-	slog.Info("initializing casbin enforcer")
+	logr.Info("initializing casbin enforcer")
 	if a.enforcer, err = NewCasbinEnforcer(casbinConfig); err != nil {
 		return nil, fmt.Errorf("failed to initialize casbin enforcer: %w", err)
 	}
 
 	// Need to refresh the cache to verify jwks is available
-	_, err = a.cache.Refresh(ctx, cfg.JwksURI)
+	_, err = cache.Refresh(ctx, oidcConfig.JwksURI)
 	if err != nil {
 		return nil, err
 	}
+
+	// Set the cache
+	a.cachedKeySet = jwk.NewCachedSet(cache, oidcConfig.JwksURI)
 
 	// Combine public routes
 	a.publicRoutes = append(a.publicRoutes, cfg.PublicRoutes...)
 	a.publicRoutes = append(a.publicRoutes, allowedPublicEndpoints[:]...)
 
-	a.oidcConfigurations[cfg.Issuer] = cfg.AuthNConfig
+	a.oidcConfiguration = cfg.AuthNConfig
+
+	// Try an register oidc issuer to wellknown service but don't return an error if it fails
+	if err := wellknownRegistration("platform_issuer", cfg.Issuer); err != nil {
+		logr.Warn("failed to register platform issuer", slog.String("error", err.Error()))
+	}
+
+	var oidcConfigMap map[string]any
+
+	// Create a map of the oidc configuration
+	oidcConfigBytes, err := json.Marshal(oidcConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := json.Unmarshal(oidcConfigBytes, &oidcConfigMap); err != nil {
+		return nil, err
+	}
+
+	if err := wellknownRegistration("idp", oidcConfigMap); err != nil {
+		logr.Warn("failed to register platform idp information", slog.String("error", err.Error()))
+	}
 
 	return a, nil
 }
@@ -158,7 +192,7 @@ func normalizeURL(o string, u *url.URL) string {
 // verifyTokenHandler is a http handler that verifies the token
 func (a Authentication) MuxHandler(handler http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if slices.ContainsFunc(a.publicRoutes, a.isPublicRoute(r.URL.Path)) {
+		if slices.ContainsFunc(a.publicRoutes, a.isPublicRoute(r.URL.Path)) { //nolint:contextcheck // There is no way to pass a context here
 			handler.ServeHTTP(w, r)
 			return
 		}
@@ -184,7 +218,7 @@ func (a Authentication) MuxHandler(handler http.Handler) http.Handler {
 		}, r.Header["Dpop"])
 
 		if err != nil {
-			slog.WarnContext(r.Context(), "failed to validate token", slog.String("error", err.Error()))
+			a.logger.WarnContext(r.Context(), "failed to validate token", slog.String("error", err.Error()))
 			http.Error(w, "unauthenticated", http.StatusUnauthorized)
 			return
 		}
@@ -203,14 +237,14 @@ func (a Authentication) MuxHandler(handler http.Handler) http.Handler {
 		}
 		if allow, err := a.enforcer.Enforce(accessTok, r.URL.Path, action); err != nil {
 			if err.Error() == "permission denied" {
-				slog.WarnContext(r.Context(), "permission denied", slog.String("azp", accessTok.Subject()), slog.String("error", err.Error()))
+				a.logger.WarnContext(r.Context(), "permission denied", slog.String("azp", accessTok.Subject()), slog.String("error", err.Error()))
 				http.Error(w, "permission denied", http.StatusForbidden)
 				return
 			}
 			http.Error(w, "internal server error", http.StatusInternalServerError)
 			return
 		} else if !allow {
-			slog.WarnContext(r.Context(), "permission denied", slog.String("azp", accessTok.Subject()))
+			a.logger.WarnContext(r.Context(), "permission denied", slog.String("azp", accessTok.Subject()))
 			http.Error(w, "permission denied", http.StatusForbidden)
 			return
 		}
@@ -222,7 +256,7 @@ func (a Authentication) MuxHandler(handler http.Handler) http.Handler {
 // UnaryServerInterceptor is a grpc interceptor that verifies the token in the metadata
 func (a Authentication) UnaryServerInterceptor(ctx context.Context, req any, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (any, error) {
 	// Allow health checks and other public routes to pass through
-	if slices.ContainsFunc(a.publicRoutes, a.isPublicRoute(info.FullMethod)) {
+	if slices.ContainsFunc(a.publicRoutes, a.isPublicRoute(info.FullMethod)) { //nolint:contextcheck // There is no way to pass a context here
 		return handler(ctx, req)
 	}
 
@@ -267,19 +301,19 @@ func (a Authentication) UnaryServerInterceptor(ctx context.Context, req any, inf
 		md["dpop"],
 	)
 	if err != nil {
-		slog.Warn("failed to validate token", slog.String("error", err.Error()))
+		a.logger.Warn("failed to validate token", slog.String("error", err.Error()))
 		return nil, status.Errorf(codes.Unauthenticated, "unauthenticated")
 	}
 
 	// Check if the token is allowed to access the resource
 	if allowed, err := a.enforcer.Enforce(token, resource, action); err != nil {
 		if err.Error() == "permission denied" {
-			slog.Warn("permission denied", slog.String("azp", token.Subject()), slog.String("error", err.Error()))
+			a.logger.Warn("permission denied", slog.String("azp", token.Subject()), slog.String("error", err.Error()))
 			return nil, status.Errorf(codes.PermissionDenied, "permission denied")
 		}
 		return nil, err
 	} else if !allowed {
-		slog.Warn("permission denied", slog.String("azp", token.Subject()))
+		a.logger.Warn("permission denied", slog.String("azp", token.Subject()))
 		return nil, status.Errorf(codes.PermissionDenied, "permission denied")
 	}
 
@@ -302,48 +336,24 @@ func (a Authentication) checkToken(ctx context.Context, authHeader []string, dpo
 		return nil, nil, fmt.Errorf("not of type bearer or dpop")
 	}
 
-	// We have to get iss from the token first to verify the signature
-	unverifiedToken, err := jwt.Parse([]byte(tokenRaw), jwt.WithVerify(false))
+	// Now we verify the token signature
+	accessToken, err := jwt.Parse([]byte(tokenRaw),
+		jwt.WithKeySet(a.cachedKeySet),
+		jwt.WithValidate(true),
+		jwt.WithIssuer(a.oidcConfiguration.Issuer),
+		jwt.WithAudience(a.oidcConfiguration.Audience),
+	)
+
 	if err != nil {
 		return nil, nil, err
 	}
 
 	// Get actor ID (sub) from unverified token for audit and add to context
-	actorID := unverifiedToken.Subject()
-	ctx = context.WithValue(ctx, audit.ActorIDContextKey, actorID)
-
-	// Get issuer from unverified token
-	issuer := unverifiedToken.Issuer()
-	if issuer == "" {
-		return nil, nil, fmt.Errorf("missing issuer")
-	}
-
-	// Get the openid configuration for the issuer
-	oidc, exists := a.oidcConfigurations[issuer]
-	if !exists {
-		validIssuers := make([]string, 0)
-		for iss := range a.oidcConfigurations {
-			validIssuers = append(validIssuers, iss)
-		}
-		return nil, nil, fmt.Errorf("invalid issuer: [%s], expected one of the configured issuers: %v", issuer, validIssuers)
-	}
-
-	// Get key set from cache that matches the jwks_uri
-	keySet, err := a.cache.Get(ctx, oidc.JwksURI)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to get jwks from cache")
-	}
-
-	// Now we verify the token signature
-	accessToken, err := jwt.Parse([]byte(tokenRaw),
-		jwt.WithKeySet(keySet),
-		jwt.WithValidate(true),
-		jwt.WithIssuer(issuer),
-		jwt.WithAudience(oidc.Audience),
-	)
-
-	if err != nil {
-		return nil, nil, err
+	// Only set the actor ID if it is not already defined
+	existingActorID := ctx.Value(sdkAudit.ActorIDContextKey)
+	if existingActorID == nil {
+		actorID := accessToken.Subject()
+		ctx = context.WithValue(ctx, sdkAudit.ActorIDContextKey, actorID)
 	}
 
 	_, tokenHasCNF := accessToken.Get("cnf")
@@ -378,6 +388,7 @@ func getContextDetails(ctx context.Context) *authContext {
 		return c
 	}
 
+	// We should probably return an error here?
 	slog.ErrorContext(ctx, "invalid authContext")
 	return nil
 }
@@ -523,12 +534,12 @@ func validateDPoP(accessToken jwt.Token, acessTokenRaw string, dpopInfo receiver
 
 func (a Authentication) isPublicRoute(path string) func(string) bool {
 	return func(route string) bool {
-		matched, err := filepath.Match(route, path)
+		matched, err := doublestar.Match(route, path)
 		if err != nil {
-			slog.Warn("error matching route", slog.String("route", route), slog.String("path", path), slog.String("error", err.Error()))
+			a.logger.Warn("error matching route", slog.String("route", route), slog.String("path", path), slog.String("error", err.Error()))
 			return false
 		}
-		slog.Debug("matching route", slog.String("route", route), slog.String("path", path), slog.Bool("matched", matched))
+		a.logger.Trace("matching route", slog.String("route", route), slog.String("path", path), slog.Bool("matched", matched))
 		return matched
 	}
 }