+++ service/internal/config/config.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
+	"reflect"
 	"strings"
 
 	"github.com/creasty/defaults"
@@ -13,10 +14,12 @@ import (
 	"github.com/opentdf/platform/service/internal/server"
 	"github.com/opentdf/platform/service/pkg/db"
 	"github.com/opentdf/platform/service/pkg/serviceregistry"
+	"github.com/opentdf/platform/service/pkg/util"
 	"github.com/spf13/viper"
 )
 
 type Config struct {
+	DevMode  bool                                     `mapstructure:"dev_mode"`
 	DB       db.Config                                `yaml:"db"`
 	OPA      opa.Config                               `yaml:"opa"`
 	Server   server.Config                            `yaml:"server"`
@@ -83,4 +86,22 @@ func LoadConfig(key string, file string) (*Config, error) {
 	}
 
 	return config, nil
+}
+
+func (c *Config) LogValue() slog.Value {
+	redactedConfig := util.RedactSensitiveData(c)
+	var values []slog.Attr
+	v := reflect.ValueOf(redactedConfig).Elem()
+	t := v.Type()
+
+	for i := 0; i < v.NumField(); i++ {
+		field := v.Field(i)
+		fieldType := t.Field(i)
+		key := fieldType.Tag.Get("yaml")
+		if key == "" {
+			key = fieldType.Name
+		}
+		values = append(values, slog.String(key, util.StructToString(field)))
+	}
+	return slog.GroupValue(values...)
 }