+++ service/internal/server/server.go
@@ -17,7 +17,9 @@ import (
 	protovalidate_middleware "github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/protovalidate"
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/realip"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
+	sdkAudit "github.com/opentdf/platform/sdk/audit"
 	"github.com/opentdf/platform/service/internal/auth"
+	"github.com/opentdf/platform/service/internal/logger"
 	"github.com/opentdf/platform/service/internal/logger/audit"
 	"github.com/opentdf/platform/service/internal/security"
 	"github.com/valyala/fasthttp/fasthttputil"
@@ -103,7 +105,7 @@ type inProcessServer struct {
 	srv *grpc.Server
 }
 
-func NewOpenTDFServer(config Config) (*OpenTDFServer, error) {
+func NewOpenTDFServer(config Config, logr *logger.Logger) (*OpenTDFServer, error) {
 	var (
 		authN *auth.Authentication
 		err   error
@@ -115,18 +117,15 @@ func NewOpenTDFServer(config Config) (*OpenTDFServer, error) {
 		authN, err = auth.NewAuthenticator(
 			context.Background(),
 			config.Auth,
+			logr,
+			config.WellKnownConfigRegister,
 		)
 		if err != nil {
 			return nil, fmt.Errorf("failed to create authentication interceptor: %w", err)
 		}
-		slog.Debug("authentication interceptor enabled")
+		logr.Debug("authentication interceptor enabled")
 	} else {
-		slog.Warn("disabling authentication. this is deprecated and will be removed. if you are using an IdP without DPoP set `enforceDPoP = false`")
-	}
-
-	// Try an register oidc issuer to wellknown service but don't return an error if it fails
-	if err := config.WellKnownConfigRegister("platform_issuer", config.Auth.Issuer); err != nil {
-		slog.Warn("failed to register platform issuer", slog.String("error", err.Error()))
+		logr.Warn("disabling authentication. this is deprecated and will be removed. if you are using an IdP without DPoP set `enforceDPoP = false`")
 	}
 
 	// Create grpc server and in process grpc server
@@ -136,7 +135,7 @@ func NewOpenTDFServer(config Config) (*OpenTDFServer, error) {
 	}
 	grpcIPCServer := &inProcessServer{
 		ln:  fasthttputil.NewInmemoryListener(),
-		srv: grpc.NewServer(),
+		srv: newGrpcInProcessServer(),
 	}
 
 	// Create http server
@@ -156,7 +155,7 @@ func NewOpenTDFServer(config Config) (*OpenTDFServer, error) {
 	}
 
 	// Create crypto provider
-	slog.Info("creating crypto provider", slog.String("type", config.CryptoProvider.Type))
+	logr.Info("creating crypto provider", slog.String("type", config.CryptoProvider.Type))
 	o.CryptoProvider, err = security.NewCryptoProvider(config.CryptoProvider)
 	if err != nil {
 		return nil, fmt.Errorf("security.NewCryptoProvider: %w", err)
@@ -233,6 +232,24 @@ func httpGrpcHandlerFunc(h http.Handler, g *grpc.Server) http.Handler {
 	})
 }
 
+func newGrpcInProcessServer() *grpc.Server {
+	var interceptors []grpc.UnaryServerInterceptor
+	var serverOptions []grpc.ServerOption
+
+	// Add audit to in process server
+	interceptors = append(interceptors, audit.ContextServerInterceptor)
+
+	// FIXME: this should probably use existing IP address instead of local?
+	// Add RealIP interceptor to in process server
+	// trustedPeers := []netip.Prefix{} // TODO: add this as a config option?
+	// headers := []string{realip.XForwardedFor, realip.XRealIp}
+	// interceptors = append(interceptors, realip.UnaryServerInterceptor(trustedPeers, headers))
+
+	// Add interceptors to server options
+	serverOptions = append(serverOptions, grpc.ChainUnaryInterceptor(interceptors...))
+	return grpc.NewServer(serverOptions...)
+}
+
 // newGrpcServer creates a new grpc server with the given config and authN interceptor
 func newGrpcServer(c Config, a *auth.Authentication) (*grpc.Server, error) {
 	var i []grpc.UnaryServerInterceptor
@@ -245,7 +262,7 @@ func newGrpcServer(c Config, a *auth.Authentication) (*grpc.Server, error) {
 	}
 
 	// Add Audit Unary Server Interceptor
-	i = append(i, audit.UnaryServerInterceptor)
+	i = append(i, audit.ContextServerInterceptor)
 
 	if c.Auth.Enabled {
 		i = append(i, a.UnaryServerInterceptor)
@@ -314,6 +331,11 @@ func (s inProcessServer) GetGrpcServer() *grpc.Server {
 }
 
 func (s inProcessServer) Conn() *grpc.ClientConn {
+	var clientInterceptors []grpc.UnaryClientInterceptor
+
+	// Add audit interceptor
+	clientInterceptors = append(clientInterceptors, sdkAudit.MetadataAddingClientInterceptor)
+
 	defaultOptions := []grpc.DialOption{
 		grpc.WithContextDialer(func(_ context.Context, _ string) (net.Conn, error) {
 			conn, err := s.ln.Dial()
@@ -323,6 +345,7 @@ func (s inProcessServer) Conn() *grpc.ClientConn {
 			return conn, nil
 		}),
 		grpc.WithTransportCredentials(insecure.NewCredentials()),
+		grpc.WithChainUnaryInterceptor(clientInterceptors...),
 	}
 
 	conn, _ := grpc.Dial("", defaultOptions...)