+++ service/kas/access/publicKey.go
@@ -19,39 +19,28 @@ import (
 const (
 	ErrCertificateEncode = Error("certificate encode error")
 	ErrPublicKeyMarshal  = Error("public key marshal error")
-	algorithmEc256       = "ec:secp256r1"
-	algorithmRSA2048     = "rsa:2048"
 )
 
 func (p Provider) lookupKid(ctx context.Context, algorithm string) (string, error) {
-	key := "unknown"
-	defaultKid := "unknown"
-	if algorithm == algorithmEc256 {
-		defaultKid = "123"
-		key = "eccertid"
+	if len(p.KASConfig.Keyring) == 0 {
+		slog.WarnContext(ctx, "no default keys found", "algorithm", algorithm)
+		return "", errors.Join(ErrConfig, status.Error(codes.NotFound, "no default keys configured"))
 	}
 
-	if p.Config == nil || p.Config.ExtraProps == nil {
-		slog.WarnContext(ctx, "using default kid", "kid", defaultKid, "algorithm", algorithm, "certid", key)
-		return defaultKid, nil
+	for _, k := range p.KASConfig.Keyring {
+		if k.Algorithm == algorithm && !k.Legacy {
+			return k.KID, nil
+		}
 	}
-
-	certid, ok := p.Config.ExtraProps[key]
-	if !ok {
-		slog.WarnContext(ctx, "using default kid", "kid", defaultKid, "algorithm", algorithm, "certid", key)
-		return defaultKid, nil
-	}
-
-	kid, ok := certid.(string)
-	if !ok {
-		slog.ErrorContext(ctx, "invalid key configuration", "kid", defaultKid, "algorithm", algorithm, "certid", key)
-		return "", ErrConfig
-	}
-	return kid, nil
+	slog.WarnContext(ctx, "no (non-legacy) key for requested algorithm", "algorithm", algorithm)
+	return "", errors.Join(ErrConfig, status.Error(codes.NotFound, "no default key for algorithm"))
 }
 
 func (p Provider) LegacyPublicKey(ctx context.Context, in *kaspb.LegacyPublicKeyRequest) (*wrapperspb.StringValue, error) {
 	algorithm := in.GetAlgorithm()
+	if algorithm == "" {
+		algorithm = security.AlgorithmRSA2048
+	}
 	var pem string
 	var err error
 	if p.CryptoProvider == nil {
@@ -63,13 +52,13 @@ func (p Provider) LegacyPublicKey(ctx context.Context, in *kaspb.LegacyPublicKey
 	}
 
 	switch algorithm {
-	case algorithmEc256:
+	case security.AlgorithmECP256R1:
 		pem, err = p.CryptoProvider.ECCertificate(kid)
 		if err != nil {
 			slog.ErrorContext(ctx, "CryptoProvider.ECPublicKey failed", "err", err)
 			return nil, errors.Join(ErrConfig, status.Error(codes.Internal, "configuration error"))
 		}
-	case algorithmRSA2048:
+	case security.AlgorithmRSA2048:
 		fallthrough
 	case "":
 		pem, err = p.CryptoProvider.RSAPublicKey(kid)
@@ -85,13 +74,16 @@ func (p Provider) LegacyPublicKey(ctx context.Context, in *kaspb.LegacyPublicKey
 
 func (p Provider) PublicKey(ctx context.Context, in *kaspb.PublicKeyRequest) (*kaspb.PublicKeyResponse, error) {
 	algorithm := in.GetAlgorithm()
+	if algorithm == "" {
+		algorithm = security.AlgorithmRSA2048
+	}
 	fmt := in.GetFmt()
 	kid, err := p.lookupKid(ctx, algorithm)
 	if err != nil {
 		return nil, err
 	}
 
-	r := func(k string, err error) (*kaspb.PublicKeyResponse, error) {
+	r := func(value, kid string, err error) (*kaspb.PublicKeyResponse, error) {
 		if errors.Is(err, security.ErrCertNotFound) {
 			slog.ErrorContext(ctx, "no key found for", "err", err, "kid", kid, "algorithm", algorithm, "fmt", fmt)
 			return nil, errors.Join(err, status.Error(codes.NotFound, "no such key"))
@@ -99,25 +91,29 @@ func (p Provider) PublicKey(ctx context.Context, in *kaspb.PublicKeyRequest) (*k
 			slog.ErrorContext(ctx, "configuration error for key lookup", "err", err, "kid", kid, "algorithm", algorithm, "fmt", fmt)
 			return nil, errors.Join(ErrConfig, status.Error(codes.Internal, "configuration error"))
 		}
-		return &kaspb.PublicKeyResponse{PublicKey: k}, nil
+		if in.GetV() == "1" {
+			slog.WarnContext(ctx, "hiding kid in public key response for legacy client", "kid", kid, "v", in.GetV())
+			return &kaspb.PublicKeyResponse{PublicKey: value}, nil
+		}
+		return &kaspb.PublicKeyResponse{PublicKey: value, Kid: kid}, nil
 	}
 
 	switch algorithm {
-	case algorithmEc256:
+	case security.AlgorithmECP256R1:
 		ecPublicKeyPem, err := p.CryptoProvider.ECPublicKey(kid)
-		return r(ecPublicKeyPem, err)
-	case algorithmRSA2048:
+		return r(ecPublicKeyPem, kid, err)
+	case security.AlgorithmRSA2048:
 		fallthrough
 	case "":
 		switch fmt {
 		case "jwk":
 			rsaPublicKeyPem, err := p.CryptoProvider.RSAPublicKeyAsJSON(kid)
-			return r(rsaPublicKeyPem, err)
+			return r(rsaPublicKeyPem, kid, err)
 		case "pkcs8":
 			fallthrough
 		case "":
 			rsaPublicKeyPem, err := p.CryptoProvider.RSAPublicKey(kid)
-			return r(rsaPublicKeyPem, err)
+			return r(rsaPublicKeyPem, kid, err)
 		}
 	}
 	return nil, status.Error(codes.NotFound, "invalid algorithm or format")