+++ service/kas/access/rewrap.go
@@ -30,7 +30,7 @@ import (
 	"github.com/opentdf/platform/service/internal/auth"
 	"github.com/opentdf/platform/service/internal/logger"
 	"github.com/opentdf/platform/service/internal/logger/audit"
-	"github.com/opentdf/platform/service/kas/tdf3"
+	"github.com/opentdf/platform/service/internal/security"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
@@ -41,13 +41,13 @@ type SignedRequestBody struct {
 }
 
 type RequestBody struct {
-	AuthToken       string         `json:"authToken"`
-	KeyAccess       tdf3.KeyAccess `json:"keyAccess"`
-	Policy          string         `json:"policy,omitempty"`
-	Algorithm       string         `json:"algorithm,omitempty"`
-	ClientPublicKey string         `json:"clientPublicKey"`
-	PublicKey       interface{}    `json:"-"`
-	SchemaVersion   string         `json:"schemaVersion,omitempty"`
+	AuthToken       string      `json:"authToken"`
+	KeyAccess       KeyAccess   `json:"keyAccess"`
+	Policy          string      `json:"policy,omitempty"`
+	Algorithm       string      `json:"algorithm,omitempty"`
+	ClientPublicKey string      `json:"clientPublicKey"`
+	PublicKey       interface{} `json:"-"`
+	SchemaVersion   string      `json:"schemaVersion,omitempty"`
 }
 
 type entityInfo struct {
@@ -289,7 +289,29 @@ func (p *Provider) Rewrap(ctx context.Context, in *kaspb.RewrapRequest) (*kaspb.
 }
 
 func (p *Provider) tdf3Rewrap(ctx context.Context, body *RequestBody, entity *entityInfo) (*kaspb.RewrapResponse, error) {
-	symmetricKey, err := p.CryptoProvider.RSADecrypt(crypto.SHA1, "UnKnown", "", body.KeyAccess.WrappedKey)
+	var kidsToCheck []string
+	if body.KeyAccess.KID != "" {
+		kidsToCheck = []string{body.KeyAccess.KID}
+	} else {
+		p.Logger.InfoContext(ctx, "kid free kao")
+		for _, k := range p.KASConfig.Keyring {
+			if k.Algorithm == security.AlgorithmRSA2048 && k.Legacy {
+				kidsToCheck = append(kidsToCheck, k.KID)
+			}
+		}
+		if len(kidsToCheck) == 0 {
+			p.Logger.WarnContext(ctx, "failure to find legacy kids for rsa")
+			return nil, err400("bad request")
+		}
+	}
+	symmetricKey, err := p.CryptoProvider.RSADecrypt(crypto.SHA1, kidsToCheck[0], "", body.KeyAccess.WrappedKey)
+	for _, kid := range kidsToCheck[1:] {
+		p.Logger.WarnContext(ctx, "continue paging through legacy KIDs for kid free kao", "err", err)
+		if err == nil {
+			break
+		}
+		symmetricKey, err = p.CryptoProvider.RSADecrypt(crypto.SHA1, kid, "", body.KeyAccess.WrappedKey)
+	}
 	if err != nil {
 		p.Logger.WarnContext(ctx, "failure to decrypt dek", "err", err)
 		return nil, err400("bad request")
@@ -308,7 +330,7 @@ func (p *Provider) tdf3Rewrap(ctx context.Context, body *RequestBody, entity *en
 		Jwt: entity.Token,
 	}
 
-	access, err := canAccess(ctx, tok, *policy, p.SDK)
+	access, err := canAccess(ctx, tok, *policy, p.SDK, *p.Logger)
 
 	// Audit the TDF3 Rewrap
 	kasPolicy := ConvertToAuditKasPolicy(*policy)
@@ -353,6 +375,13 @@ func (p *Provider) tdf3Rewrap(ctx context.Context, body *RequestBody, entity *en
 }
 
 func (p *Provider) nanoTDFRewrap(ctx context.Context, body *RequestBody, entity *entityInfo) (*kaspb.RewrapResponse, error) {
+	// TODO Lookup KID from request content
+	// Should this be in the locator or somewhere else?
+	kid, err := p.lookupKid(ctx, security.AlgorithmECP256R1)
+	if err != nil {
+		p.Logger.WarnContext(ctx, "failure to find default kid for ec", "err", err)
+		return nil, err400("bad request")
+	}
 	headerReader := bytes.NewReader(body.KeyAccess.Header)
 
 	header, _, err := sdk.NewNanoTDFHeaderFromReader(headerReader)
@@ -365,7 +394,7 @@ func (p *Provider) nanoTDFRewrap(ctx context.Context, body *RequestBody, entity 
 		return nil, fmt.Errorf("ECCurve failed: %w", err)
 	}
 
-	symmetricKey, err := p.CryptoProvider.GenerateNanoTDFSymmetricKey(header.EphemeralKey, ecCurve)
+	symmetricKey, err := p.CryptoProvider.GenerateNanoTDFSymmetricKey(kid, header.EphemeralKey, ecCurve)
 	if err != nil {
 		return nil, fmt.Errorf("failed to generate symmetric key: %w", err)
 	}
@@ -392,7 +421,7 @@ func (p *Provider) nanoTDFRewrap(ctx context.Context, body *RequestBody, entity 
 		Jwt: entity.Token,
 	}
 
-	access, err := canAccess(ctx, tok, *policy, p.SDK)
+	access, err := canAccess(ctx, tok, *policy, p.SDK, *p.Logger)
 
 	// Audit the rewrap
 	kasPolicy := ConvertToAuditKasPolicy(*policy)